Audit App hibernate changes

+++ src/main/java/com/heights/auditapp/service/AuditFocusService.java
@@ -2,7 +2,6 @@ package com.heights.auditapp.service;
 
 import com.heights.auditapp.model.AuditFocusEntity;
 
-import java.math.BigInteger;
 
-public interface AuditFocusService extends GenericService<AuditFocusEntity, BigInteger> {
+public interface AuditFocusService extends GenericService<AuditFocusEntity, Long> {
 }
