+++ src/main/java/com/heights/auditapp/service/impl/AuditFocusServiceImpl.java
@@ -9,7 +9,6 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
-import java.math.BigInteger;
 import java.util.List;
 import java.util.Optional;
 
@@ -33,12 +32,12 @@ public class AuditFocusServiceImpl implements AuditFocusService {
     }
 
     @Override
-    public void deleteById(BigInteger id) {
+    public void deleteById(Long id) {
         repository.deleteById(id);
     }
 
     @Override
-    public Optional<AuditFocusEntity> findById(BigInteger id) {
+    public Optional<AuditFocusEntity> findById(Long id) {
         return repository.findById(id);
     }
 
@@ -55,7 +54,7 @@ public class AuditFocusServiceImpl implements AuditFocusService {
     }
 
     @Override
-    public AuditFocusEntity update(AuditFocusEntity entity, BigInteger id) {
+    public AuditFocusEntity update(AuditFocusEntity entity, Long id) {
         Optional<AuditFocusEntity> optional = findById(id);
         if (optional.isPresent()) {
             return save(entity);
